Using Spring Boot Security Auth to get current (bean) user

+++ amexproject2/src/main/java/com/example/amexproject2/service/PostServiceImpl.java
@@ -1,15 +1,23 @@
 package com.example.amexproject2.service;
 
+import com.example.amexproject2.controller.SecurityController;
 import com.example.amexproject2.model.Post;
 import com.example.amexproject2.model.User;
 import com.example.amexproject2.repository.PostRepository;
 import com.example.amexproject2.repository.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
+import org.springframework.security.authentication.AnonymousAuthenticationToken;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
 
 @Service
 public class PostServiceImpl implements PostService {
+
+    @Autowired
+    SecurityController securityController;
 
     @Autowired
     private PostRepository postRepository;
@@ -17,19 +25,37 @@ public class PostServiceImpl implements PostService {
     @Autowired
     private UserRepository userRepository;
 
-
     @Override
     public Post createPost(String username,
                            Post newPost) {
         User userWhoPosts = userRepository.findByUsername(username);
-        newPost.setUser(userWhoPosts);
-        return postRepository.save(newPost);
+        if (userWhoPosts != null) {
+            newPost.setUser(userWhoPosts);
+            return postRepository.save(newPost);
+        }
+        System.out.println(HttpStatus.BAD_REQUEST);
+       return null;
     }
 
     @Override
-    public HttpStatus deleteById(Long postId){
-        postRepository.deleteById(postId);
-        return HttpStatus.OK;
+    public HttpStatus deleteById(Long postId) {
+
+        //        String currentUsername = null;
+//
+//        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+//        if (!(authentication instanceof AnonymousAuthenticationToken)) {
+//            currentUsername = authentication.getName();
+//        }
+
+        String currentUsername = securityController.getCurrentUsername();
+
+
+        if (postRepository.findById(postId).get().getUser().getUsername().equals(currentUsername)) {
+            postRepository.deleteById(postId);
+            return HttpStatus.OK;
+        } else {
+            return HttpStatus.BAD_REQUEST;
+        }
     }
 
     public Post getPostById(Long postId) {
